pyproject.toml
[tool.poetry]
name = "poetryneomodelsfastapitest"
version = "0.1.0"
description = "A FastAPI project using Poetry and Neomodels"
authors = ["Orthodoxos Frangoullis"]
repository = "https://github.com/orthosf/Neo4j-neomodels-FastAPI-Docker-Poetry"
homepage = "https://github.com/orthosf/Neo4j-neomodels-FastAPI-Docker-Poetry"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.111.1"
uvicorn = "0.30.1"
django = "^3.2" 
annotated-types = "0.7.0"
anyio = "4.4.0"
bcrypt = "4.1.3"
build = "1.2.1"
CacheControl = "0.14.0"
certifi = "2024.7.4"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
click = "8.1.7"
crashtest = "0.4.1"
cryptography = "42.0.8"
distlib = "0.3.8"
dnspython = "2.6.1"
dulwich = "0.21.7"
email-validator = "2.2.0"
fastapi-cli = "0.0.4"
fastjsonschema = "2.20.0"
filelock = "3.15.4"
greenlet = "3.0.3"
h11 = "0.14.0"
httpcore = "1.0.5"
httptools = "0.6.1"
httpx = "0.27.0"
idna = "3.7"
iniconfig = "2.0.0"
installer = "0.7.0"
jeepney = "0.8.0"
jinja2 = "3.1.4"
keyring = "24.3.1"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
mdurl = "0.1.2"
more-itertools = "10.3.0"
msgpack = "1.0.8"
neo4j = "5.19.0"
neomodel = "5.3.2"
packaging = "24.1"
passlib = "1.7.4"
pexpect = "4.9.0"
pkginfo = "1.11.1"
platformdirs = "4.2.2"
pluggy = "1.5.0"
poetry = "1.8.3"
poetry-core = "1.9.0"
poetry-plugin-export = "1.8.0"
ptyprocess = "0.7.0"
pycparser = "2.22"
pydantic = "2.8.2"
pydantic-core = "2.20.1"
pygments = "2.18.0"
pyproject-hooks = "1.1.0"
pytest = "8.3.1"
pytest-mock = "3.14.0"
python-dotenv = "1.0.1"
python-multipart = "0.0.9"
pytz = "2024.1"
pyyaml = "6.0.1"
rapidfuzz = "3.9.4"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
rich = "13.7.1"
secretstorage = "3.3.3"
setuptools = "70.3.0"
shellingham = "1.5.4"
sniffio = "1.3.1"
sqlalchemy = "2.0.31"
sqlmodel = "0.0.20"
starlette = "0.37.2"
tomlkit = "0.13.0"
trove-classifiers = "2024.7.2"
typer = "0.12.3"
typing-extensions = "4.12.2"
urllib3 = "2.2.2"
uuid = "1.30"
uvloop = "0.19.0"
virtualenv = "20.26.3"
watchfiles = "0.22.0"
websockets = "12.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.1"
pytest-mock = "^3.14.0"
nox = "^2023.3.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"